+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/factory/NDArrayFactory.java
@@ -31,7 +31,6 @@ import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.api.rng.distribution.Distribution;
 
 import java.io.File;
-import java.io.InputStream;
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
@@ -1928,21 +1927,21 @@ public interface NDArrayFactory {
 
     // =========== Sparse methods ===========
 
-    INDArray createSparseCSR(double[] data, int[] columns, int[] pointerB, int[] pointerE, int[] shape);
+    INDArray createSparseCSR(double[] data, int[] columns, int[] pointerB, int[] pointerE, long[] shape);
 
-    INDArray createSparseCSR(float[] data, int[] columns, int[] pointerB, int[] pointerE, int[] shape);
+    INDArray createSparseCSR(float[] data, int[] columns, int[] pointerB, int[] pointerE, long[] shape);
 
-    INDArray createSparseCSR(DataBuffer data, int[] columns, int[] pointerB, int[] pointerE, int[] shape);
+    INDArray createSparseCSR(DataBuffer data, int[] columns, int[] pointerB, int[] pointerE, long[] shape);
 
-    INDArray createSparseCOO(double[] values, int[][] indices, int[] shape);
+    INDArray createSparseCOO(double[] values, int[][] indices, long[] shape);
 
-    INDArray createSparseCOO(float[] values, int[][] indices, int[] shape);
+    INDArray createSparseCOO(float[] values, int[][] indices, long[] shape);
 
-    INDArray createSparseCOO(DataBuffer values, DataBuffer indices, int[] shape);
+    INDArray createSparseCOO(DataBuffer values, DataBuffer indices, long[] shape);
 
-    INDArray createSparseCOO(DataBuffer values, DataBuffer indices, DataBuffer sparseInformation, int[] shape);
+    INDArray createSparseCOO(DataBuffer values, DataBuffer indices, DataBuffer sparseInformation, long[] shape);
 
     INDArray createSparseCOO(DataBuffer values, DataBuffer indices, long[] sparseOffsets, int[] flags,
-                    int[] hiddenDimensions, int underlyingRank, int[] shape);
+                             int[] hiddenDimensions, int underlyingRank, long[] shape);
 
 }
